completed implementation of project page

+++ src/main/java/com/resource/manager/resource/repository/ProjectRepository.java
@@ -5,6 +5,6 @@ import com.resource.manager.resource.entity.Project;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectRepository extends BaseRepository<Project, Long> {
+public interface ProjectRepository extends BaseRepository<Project, Integer> {
 
 }
